requirement.txt
rich
typer[all]
GitPython
watchdog
